Making twitter ID a big int (INT8)

+++ db/migrate/20100109043057_create_status_updates.rb
@@ -2,7 +2,7 @@ class CreateStatusUpdates < ActiveRecord::Migration
   def self.up
     create_table :status_updates do |t|
       t.integer :rubyist_id
-      t.integer :twitter_id
+      t.int8 :twitter_id
       t.text :original_text
       t.text :translated_text
